Add PayPalWebVaultRequest to project.
