+++ PayPalWebPayments/src/main/java/com/paypal/android/paypalwebpayments/PayPalWebVaultRequest.kt
@@ -1,0 +1,6 @@
+package com.paypal.android.paypalwebpayments
+
+data class PayPalWebVaultRequest(
+    val setupTokenId: String,
+    val approveVaultHref: String
+)
